Add a confirmation

+++ cmd/certificates.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/AlecAivazis/survey/v2"
 	"github.com/logrusorgru/aurora"
 	"github.com/spf13/cobra"
 	"github.com/superfly/flyctl/cmd/presenters"
@@ -22,6 +23,7 @@ func newCertificatesCommand() *Command {
 	add.Command.Args = cobra.ExactArgs(1)
 	delete := BuildCommand(cmd, runCertDelete, "delete <hostname>", "delete new certificate", os.Stdout, true, requireAppName)
 	delete.Command.Args = cobra.ExactArgs(1)
+	delete.AddBoolFlag(BoolFlagOpts{Name: "yes", Shorthand: "y", Description: "accept all confirmations"})
 	show := BuildCommand(cmd, runCertShow, "show <hostname>", "show detailed certificate info", os.Stdout, true, requireAppName)
 	show.Command.Args = cobra.ExactArgs(1)
 	check := BuildCommand(cmd, runCertCheck, "check <hostname>", "check dns configuration", os.Stdout, true, requireAppName)
@@ -74,6 +76,18 @@ func runCertAdd(ctx *CmdContext) error {
 
 func runCertDelete(ctx *CmdContext) error {
 	hostname := ctx.Args[0]
+
+	if !ctx.Config.GetBool("yes") {
+		confirm := false
+		prompt := &survey.Confirm{
+			Message: fmt.Sprintf("Remove certificate %s from app %s?", hostname, ctx.AppName()),
+		}
+		survey.AskOne(prompt, &confirm)
+
+		if !confirm {
+			return nil
+		}
+	}
 
 	cert, err := ctx.FlyClient.DeleteCertificate(ctx.AppName(), hostname)
 	if err != nil {
